+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## DEV
 
 - Clean-up Platinum Legion Walker Foundry builds
+- Added Uber Turtle, a personality which puts more emphasis on defending its expansions
 
 ## v4.62.0 - 2020-04-25